backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy==2.0.*
alembic==1.14.*
psycopg2-binary==2.9.*
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
bcrypt==4.0.*
python-multipart==0.0.*
workalendar==17.*
openpyxl==3.1.*
pydantic==2.*
email-validator==2.*
pydantic-settings==2.*
python-dotenv==1.0.*
slowapi==0.1.*
pytest==8.*
pytest-asyncio==0.*
httpx==0.*
